Landing + Dynamics laboratory
vision_landing:
- Integrated SmartLanding algorithm framework (for computing landing trajectory and analyzing drone dynamics, generate real time graphs, etc)
- Added "stats" class to measure drone dynamics
- Use drone's yaw (sent together with the image)
- Added state machine to perform dynamic tests and perform landing.
- Fixed signal handling for clean shutdown.
- Enhanced console messaging

track_targets.cpp:
- Moved apriltag.* sources (changed directory structure)
- Replaced piped-buffer for raw tcp video receiver

config.json:
- Updated configuration for detecting new mavlink-camera-simulator's AprilTag
