+++ config.json
@@ -2,21 +2,21 @@
 	"markers": [
 		{
 			"id": 199,
-			"size": 0.21180257,
-			"offsetX": 0.476052,
-			"offsetY": 0
+			"size": 0.135,
+			"offsetX": 0.026349,
+            "offsetY": 0.141789
 		},
 		{
 			"id": 118,
-			"size": 0.066566523,
-			"offsetX": -0.088755,
-			"offsetY": -0.003026
+			"size": 0.03,
+			"offsetX": 0.046997,
+			"offsetY": -0.013023
 		},
 		{
 			"id": 85,
-			"size": 0.030257510,
-			"offsetX": 0.018154504,
-			"offsetY": 0.072618025
+			"size": 0.014,
+			"offsetX": 0.0,
+			"offsetY": 0.0
 		}
 	]
 }
